activedir_login/from_pano/app_azure/requirements.txt
# Flask 1.* used to work, until their upstream package brought in a breaking change https://github.com/pallets/markupsafe/issues/286
Flask==2.0.1

werkzeug==2.1.2
flask-session>=0.3.2,<0.5
requests>=2,<3
msal>=1.7,<2
flask-cors==3.0.10

azure-core==1.1.1
azure-cosmos==4.1.0
certifi==2019.11.28
chardet==3.0.4
idna==2.8
#requests==2.22.0
six==1.13.0
urllib3==1.25.7
virtualenv==16.7.9
virtualenv-clone==0.5.3

# cachelib==0.1  # Only need this if you are running Python 2
# Note: This sample does NOT directly depend on cachelib.
# It is an indirect dependency of flask-session.
# Cachelib 0.1.1 no longer supports Python 2
# (see also https://github.com/pallets/cachelib/issues/14)
# So, if you still need to run your app in Python 2,
# your workaround is to pin cachelib to its older version 0.1,
# but keep in mind it contains a known bug https://github.com/pallets/cachelib/pull/12
